# api/swagger.yaml
basePath: /api/v1
definitions:
  booking.BookingResponse:
    properties:
      contact_email:
        description: Contact email
        example: user@example.com
        type: string
      created_at:
        description: Creation timestamp in RFC 3339
        example: "2025-07-15T10:00:00Z"
        type: string
      currency:
        description: Currency code
        example: RUB
        type: string
      expires_at:
        description: Reservation expiry time in RFC 3339 (for pending bookings)
        example: "2025-07-15T10:15:00Z"
        type: string
      flight_id:
        description: Flight ID
        example: 1
        type: integer
      id:
        description: Booking UUID
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      ref_code:
        description: Short reference code
        example: A1B2C3
        type: string
      status:
        description: Booking status
        enum:
        - pending
        - paid
        - cancelled
        example: pending
        type: string
      tickets:
        description: Tickets
        items:
          $ref: '#/definitions/booking.TicketResponse'
        type: array
      total_price_cents:
        description: Total price in cents
        example: 1000000
        type: integer
      user_id:
        description: Owner user ID
        example: 1
        type: integer
    type: object
  booking.CreateBookingRequest:
    properties:
      contact_email:
        description: Contact email for booking notifications
        example: user@example.com
        type: string
      flight_id:
        description: Flight ID
        example: 1
        type: integer
      passengers:
        description: List of passengers
        items:
          $ref: '#/definitions/booking.PassengerInput'
        minItems: 1
        type: array
    required:
    - contact_email
    - flight_id
    - passengers
    type: object
  booking.CreateBookingResponse:
    properties:
      booking_id:
        description: Booking UUID
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      expires_at:
        description: Reservation expiry time in RFC 3339
        example: "2025-07-15T10:15:00Z"
        type: string
      ref_code:
        description: Short reference code for the passenger
        example: A1B2C3
        type: string
    type: object
  booking.PassengerInput:
    properties:
      document_number:
        description: Travel document number
        example: "7712345678"
        type: string
      first_name:
        description: Legal first name as in travel document
        example: Ivan
        type: string
      last_name:
        description: Legal last name as in travel document
        example: Ivanov
        type: string
      seat_class:
        description: 'Seat class: economy, business, first'
        enum:
        - economy
        - business
        - first
        example: economy
        type: string
      seat_number:
        description: Selected seat number
        example: 12A
        type: string
    required:
    - document_number
    - first_name
    - last_name
    - seat_class
    - seat_number
    type: object
  booking.TicketResponse:
    properties:
      id:
        description: Ticket UUID
        example: 660e8400-e29b-41d4-a716-446655440000
        type: string
      passenger_doc_num:
        description: Passenger document number
        example: "7712345678"
        type: string
      passenger_first_name:
        description: Passenger first name
        example: Ivan
        type: string
      passenger_last_name:
        description: Passenger last name
        example: Ivanov
        type: string
      price_cents:
        description: Ticket price in cents
        example: 500000
        type: integer
      seat_class:
        description: Seat class
        enum:
        - economy
        - business
        - first
        example: economy
        type: string
      seat_number:
        description: Seat number
        example: 12A
        type: string
      status:
        description: Ticket status
        enum:
        - reserved
        - issued
        example: reserved
        type: string
      ticket_number:
        description: Ticket number (assigned after issuance)
        example: SU1234567890
        type: string
    type: object
  common.ErrorResponse:
    description: Standard error response
    properties:
      message:
        description: Human-readable error message
        example: 'invalid input body: field email is required'
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Gateway API for flight booking service
  title: Flyte API
  version: "1.0"
paths:
...
  /bookings:
    get:
      description: Returns all bookings for the authenticated user with pagination.
      parameters:
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of bookings
          schema:
            items:
              $ref: '#/definitions/booking.BookingResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user bookings
      tags:
      - bookings
    post:
      consumes:
      - application/json
      description: Creates a new booking with one or more passengers and reserves
        their seats.
      parameters:
      - description: Booking data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/booking.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Booking created
          schema:
            $ref: '#/definitions/booking.CreateBookingResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "409":
          description: Seat already taken
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a booking
      tags:
      - bookings
  /bookings/{id}:
    get:
      description: Returns full details for a booking. Only the booking owner can
        access it.
      parameters:
      - description: Booking ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Booking details
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "400":
          description: Invalid booking ID
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden — not the booking owner
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Booking not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get booking details
      tags:
      - bookings
  /bookings/{id}/cancel:
    post:
      description: Cancels a booking and releases reserved seats. Only the booking
        owner can cancel.
      parameters:
      - description: Booking ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Booking cancelled
        "400":
          description: Invalid booking ID
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden — not the booking owner
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Booking not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel a booking
      tags:
      - bookings
  /bookings/ref/{ref}:
    get:
      description: Returns full details for a booking by its short reference code.
        Only the booking owner can access it.
      parameters:
      - description: Reference code
        in: path
        name: ref
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Booking details
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "403":
          description: Forbidden — not the booking owner
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Booking not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get booking by reference code
      tags:
      - bookings
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"